feat: refactor service imports and create starwars service for character fetching
