+++ src/configs/index.ts
@@ -1,4 +1,3 @@
 export * from "./axios.config";
 export * from "./swr.config";
 export * from "./query-client.config";
-export * from "./query-function.config";
